+++ django_base64field/tests/test_base64field.py
@@ -1,44 +1,12 @@
-from django.db import models
+import uuid
 from django.test import TestCase
 from django.utils.baseconv import base64
-from django_base64field.fields import Base64Field
-
-class Planet(models.Model):
-    ek = Base64Field()
-    name = models.CharField(
-        default='Fucker',
-        max_length=103
-    )
-
-
-class Continent(models.Model):
-    ek = Base64Field()
-    name = models.CharField(
-        default='Suckers!',
-        max_length=13
-    )
-    planet = models.ForeignKey(Planet, to_field='ek')
-
-
-class Helper(models.Model):
-    """
-    base64 encoded value won't be available at first time creation.
-    It can ve accessible by getting the object from database after creation
-    mean when it get saved completely, But what if we don't want to get our base64
-    encoded key from our sweet model by retrieving it again from database?
-
-    It's easy, efficient, holly and molly!
-    """
-    ek = Base64Field()
-
-
-    def _ek(self):
-        if self.ek: return self.ek
-
-        if (not self.ek) and (self.pk):
-            return base64.encode(self.pk)
-
-        return self.ek
+from django_base64field.tests.models import (
+    Planet,
+    Continent,
+    Helper,
+    CustomReceiver
+)
 
 
 class TestBase64Field(TestCase):
@@ -108,7 +76,6 @@ class TestBase64Field(TestCase):
         continent_pk = continent.pk
         same_continent_but_fresh = Continent.objects.get(pk=continent_pk)
 
-
         self.assertNotEqual(same_continent_but_fresh.ek, continent.ek)
         self.assertEqual(same_continent_but_fresh.ek, base64.encode(continent_pk))
 
@@ -131,3 +98,16 @@ class TestBase64Field(TestCase):
         self.assertIsNotNone(hell._ek())
         self.assertNotIn(hell._ek(), ['', None])
 
+    def test_field_with_custom_receiver_method(self):
+        """
+         Should use a different receiver method for generating a custom
+        base64field. For this test `CustomReceiver` model will be used which
+        it will use `django_baset64field.tests.receivers:custom_receiver` as a
+        custom receiver for `Base64field`.
+        """
+        obj = CustomReceiver.objects.create()
+        refreshed_obj = CustomReceiver.objects.get(pk=obj.pk)
+
+        self.assertIsNotNone(refreshed_obj.youyouid)
+        self.assertIsInstance(str(refreshed_obj.youyouid), str)
+        self.assertIsInstance(uuid.UUID(refreshed_obj.youyouid), uuid.UUID)